Mute the home arrow when home is not set.
When the home position is not set, the home arrow currently acts like a compass needle, and then silently transforms into a real home bearing when there is a home lock. This patch makes it so the home arrow does not move when home is not set, and gives it muted grey appearance by only showing the outline. Then, when home is set, the arrow begins to move and is filled in with white as it normal appears.

The distance to home is also shown as "H -M" instead of "H 0M" when home position is unset to make it clear the distance to home is unknowable since there is no home position yet. Then, when home is set, it reads "H XM" where X is the current number of meters to home.

Some of the code already effectively returned 0.0f for bearing/distance when home wasn't set, but I made it so the functions explicitly handle this case to make it clearer what's going on.